SI-9369 Fix pattern matcher warnings for diamond shaped inheritance.
A previous optimization (d44a86f432a7f9ca250b014acdeab02ac9f2c304) for
pattern matcher exhaustivity checks used a smarter encoding to ensure
that the scrutinee can be equal to one child only.
However, in case of traits between the root and leave type, a child can
be of several types and these types should not be in a mutually exclusive
group.

E.g.,

```Scala
  trait Tree

  sealed abstract class Prop

  trait Simple extends Prop

  case class Atom(tree: Tree) extends Prop with Simple

  case class Not(prop: Prop) extends Prop with Simple
```

The proposed approach orders the class hierarchy first
according to how many sub-classes exist for each class:

           +--------+
2          |  Prop  |
           +--------+
--           |  |  |
       +-----+  |  +-----+
       |        v        |
       |   +---------+   |
1      |   | Simple  |   |
       |   +---------+   |
       |      |   |      |
--     |   +--+   +--+   |
	   |   |         |   |
       v   v         v   v
     +-------+    +--------+
0    |  Atom |    |   Not  |
     +-------+    +--------+

In a second step, only the childs with the same number of sub-classes
are placed in mutually exclusive groups. This ensures that `Atom` and `Simple`
are not put in the same group anymore.